// web-app/customer/schemas/api-types.ts
export interface components {
  schemas: {
// ...
    Appointment: {
      id?: number;
      symptoms: string;
      symptom_category: string;
      symptoms_duration_seconds: number;
      priority?: number | null;
      state?: string;
      patient_id?: number;
      practice_id?: number;
      documents?: {
          id?: number;
          download_url: string;
          /** Format: date-time */
          uploaded_at?: string;
        }[];
      /** Format: date-time */
      created_at?: string;
      /** Format: date-time */
      updated_at?: string;
      assigned_to_id?: number | null;
      logs?: {
          id?: number;
          from_state?: string;
          to_state: string;
          triggered_by_id: number;
          triggered_by?: {
            readonly id?: number;
            readonly first_name: string;
            readonly last_name: string;
            readonly full_name?: string;
            /** Format: email */
            readonly email: string;
          };
          /** Format: date-time */
          created_at?: string;
          /** Format: date-time */
          transition_away_at?: string;
          transition_delta?: number;
          appointment_id: number;
        }[];
      comments?: {
          id?: number;
          user_id?: number;
          user?: {
            readonly id?: number;
            readonly first_name: string;
            readonly last_name: string;
            readonly full_name?: string;
            /** Format: email */
            readonly email: string;
          };
          /** Format: date-time */
          created_at?: string;
          /** Format: date-time */
          updated_at?: string;
          comment: string;
          appointment_id: number;
        }[];
      patient?: {
        address_line_1: string;
        address_line_2: string;
        city: string;
        state: string;
        zip_code: string;
        country: string;
        latitude?: number;
        longitude?: number;
        id?: number;
        first_name: string;
        last_name: string;
        full_name?: string;
        /** Format: email */
        email: string;
        phone: string;
        /** Format: date */
        date_of_birth: string;
        gender: string;
        health_care_number?: string;
        documents?: {
            id?: number;
            download_url: string;
            state?: string;
            /** Format: date-time */
            uploaded_at?: string;
            is_id: boolean;
            is_proof_of_address: boolean;
          }[];
      };
      assigned_to?: {
        id?: number;
        first_name: string;
        last_name: string;
        full_name?: string;
        /** Format: email */
        email: string;
      } | null;
      assign_history?: {
          id?: number;
          /** Format: date-time */
          created_at?: string;
          triggered_by?: {
            readonly id?: number;
            readonly first_name: string;
            readonly last_name: string;
            readonly full_name?: string;
            /** Format: email */
            readonly email: string;
          };
          from_user?: {
            readonly id?: number;
            readonly first_name: string;
            readonly last_name: string;
            readonly full_name?: string;
            /** Format: email */
            readonly email: string;
          };
          to_user?: {
            readonly id?: number;
            readonly first_name: string;
            readonly last_name: string;
            readonly full_name?: string;
            /** Format: email */
            readonly email: string;
          };
          appointment_id: number;
        }[];
      viewed_logs?: {
          id?: number;
          viewed_by?: {
            readonly id?: number;
            readonly first_name: string;
            readonly last_name: string;
            readonly full_name?: string;
            /** Format: email */
            readonly email: string;
          };
          /** Format: date-time */
          created_at?: string;
          appointment_id: number;
        }[];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}